files and text modules reorganized and expanded
