+++ dhelp/files/__init__.py
@@ -1,0 +1,4 @@
+#!/usr/bin/python
+
+from .txt import TextFile, TextFolder
+from .csv import CSVFile
