+++ dhelp/text/__init__.py
@@ -1,0 +1,4 @@
+#!/usr/bin/python
+
+from .nltk import EnglishText
+from .cltk import LatinText, AncientGreekText
